Update Traffic Analytics app structure
ref PROD-216
