+++ apps/stats/src/components/layout/Header.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import {H1} from '@tryghost/shade';
+
+interface HeaderProps extends React.HTMLAttributes<HTMLElement> {
+}
+
+const Header:React.FC<HeaderProps> = ({...props}) => {
+    return (
+        <header className='sticky top-0 z-50 -mx-8 mb-4 bg-white/85 backdrop-blur-md dark:bg-black'>
+            <div className='relative flex h-[102px] items-center justify-between gap-5 px-8 before:absolute before:inset-x-8 before:bottom-0 before:block before:border-b before:border-gray-200 before:content-[""] before:dark:border-gray-950'>
+                <H1>{props.children}</H1>
+            </div>
+        </header>
+
+    );
+};
+
+export default Header;
