+++ apps/stats/src/components/layout/index.ts
@@ -1,1 +1,1 @@
-export {default} from './Layout';
+export {default} from './MainLayout';
