+++ apps/stats/src/routes.tsx
@@ -1,4 +1,6 @@
-import Overview from './views/Overview';
+import Locations from './views/Stats/Locations';
+import Sources from './views/Stats/Sources';
+import Web from './views/Stats/Web';
 import {RouteObject} from '@tryghost/admin-x-framework';
 
 export const APP_ROUTE_PREFIX = '/stats-x';
@@ -7,6 +9,18 @@ export const routes: RouteObject[] = [
     {
         path: '',
         index: true,
-        element: <Overview />
+        element: <Web />
+    },
+    {
+        path: '/web/',
+        element: <Web />
+    },
+    {
+        path: '/sources/',
+        element: <Sources />
+    },
+    {
+        path: '/locations/',
+        element: <Locations />
     }
 ];
