+++ apps/stats/src/views/Stats/Sources.tsx
@@ -1,0 +1,16 @@
+import Header from '@src/components/layout/Header';
+import React from 'react';
+import StatsLayout from './layout/StatsLayout';
+
+const Sources:React.FC = () => {
+    return (
+        <StatsLayout>
+            <Header>Sources</Header>
+            <div>
+                Sources
+            </div>
+        </StatsLayout>
+    );
+};
+
+export default Sources;
