+++ apps/stats/src/views/Stats/index.ts
